refactor: utils 폴더 구조를 역할별로 정리
